stage_2/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.6
python-dotenv==1.0.0